Refactor transform method on SVGAttributedFigures
