+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/figures/SVGAttributedFigure.java
@@ -17,6 +17,8 @@ import javax.swing.*;
 import org.jhotdraw.draw.*;
 import static org.jhotdraw.draw.AttributeKeys.STROKE_WIDTH;
 import static org.jhotdraw.draw.AttributeKeys.TRANSFORM;
+
+import org.jhotdraw.samples.svg.Gradient;
 import org.jhotdraw.samples.svg.SVGAttributeKeys;
 import static org.jhotdraw.samples.svg.SVGAttributeKeys.*;
 import org.jhotdraw.util.*;
@@ -95,6 +97,44 @@ public abstract class SVGAttributedFigure extends AbstractAttributedFigure {
         if (get(TRANSFORM) != null) {
             g.setTransform(savedTransform);
         }
+    }
+
+    /**
+     * Transforms the figure.
+     *
+     * @param tx The transformation.
+     */
+    @Override
+    public void transform(AffineTransform tx) {
+        if (get(TRANSFORM) != null
+                || (tx.getType() & (AffineTransform.TYPE_TRANSLATION)) != tx.getType()) {
+            if (get(TRANSFORM) == null) {
+                set(TRANSFORM, (AffineTransform) tx.clone());
+            } else {
+                AffineTransform t = TRANSFORM.getClone(this);
+                t.preConcatenate(tx);
+                set(TRANSFORM, t);
+            }
+        } else {
+            Point2D.Double anchor = getStartPoint();
+            Point2D.Double lead = getEndPoint();
+            setBounds(
+                    (Point2D.Double) tx.transform(anchor, anchor),
+                    (Point2D.Double) tx.transform(lead, lead));
+            if (get(FILL_GRADIENT) != null
+                    && !get(FILL_GRADIENT).isRelativeToFigureBounds()) {
+                Gradient g = FILL_GRADIENT.getClone(this);
+                g.transform(tx);
+                set(FILL_GRADIENT, g);
+            }
+            if (get(STROKE_GRADIENT) != null
+                    && !get(STROKE_GRADIENT).isRelativeToFigureBounds()) {
+                Gradient g = STROKE_GRADIENT.getClone(this);
+                g.transform(tx);
+                set(STROKE_GRADIENT, g);
+            }
+        }
+        invalidate();
     }
 
     @Override
